properly split contract and program routing
